[김수찬] - 유저페이지 리뷰 불러오기 구현완료

+++ src/main/java/com/green/glampick/dto/request/user/PostReviewRequestDto.java
@@ -18,11 +18,9 @@ public class PostReviewRequestDto {
     @JsonIgnore private long reviewId;
     private long glamp_id;
     private String reviewContent;
-    private long reviewStarPoint;
+    private int reviewStarPoint;
     private long roomId;
+
     private List<MultipartFile> ReviewImageFiles = new ArrayList<>();
-
-
-
 
 }
